+++ src/Sandbox.stories.tsx
@@ -1,9 +1,10 @@
 import View from "components/View";
 import Image from "components/Image";
-import NumberField from "components/NumberField";
+import Table from "components/Table";
 
 export default {
 	title: "Sandbox",
+	chromatic: { disableSnapshot: true },
 };
 
 const Preview: React.FC<{ children: React.ReactNode }> = (props) => {
@@ -21,7 +22,38 @@ const Preview: React.FC<{ children: React.ReactNode }> = (props) => {
 export const preview = () => {
 	return (
 		<Preview>
-			<NumberField name="" decreaseAriaLabel="" increaseAriaLabel="" />
+			<Table>
+				<Table.Head>
+					<Table.Row>
+						<Table.Heading paddingInline={0} width="120px">
+							Column 1
+						</Table.Heading>
+						<Table.Heading paddingInline={0} width="120px">
+							Column 2
+						</Table.Heading>
+						<Table.Heading paddingInline={0} width="120px">
+							Column 3
+						</Table.Heading>
+						<Table.Heading paddingInline={0} width="120px">
+							Column 4
+						</Table.Heading>
+					</Table.Row>
+				</Table.Head>
+				<Table.Body>
+					<Table.Row>
+						<Table.Cell paddingInline={0}>Cell 1</Table.Cell>
+						<Table.Cell paddingInline={0}>Cell 2</Table.Cell>
+						<Table.Cell paddingInline={0}>Cell 3</Table.Cell>
+						<Table.Cell paddingInline={0}>Cell 4</Table.Cell>
+					</Table.Row>
+					<Table.Row>
+						<Table.Cell paddingInline={0}>Cell 1</Table.Cell>
+						<Table.Cell paddingInline={0}>Cell 2</Table.Cell>
+						<Table.Cell paddingInline={0}>Cell 3</Table.Cell>
+						<Table.Cell paddingInline={0}>Cell 4</Table.Cell>
+					</Table.Row>
+				</Table.Body>
+			</Table>
 		</Preview>
 	);
 };